refactor(gateway): extract obligation wiring (#2996 P8 PR-C2)
Move the obligation open/close/cancel helpers + the idle sweep VERBATIM from
gateway.ts into telegram-plugin/gateway/obligation-wiring.ts:
closeObligationOnSubstantiveReply, openObligationFromInbound,
cancelInterruptedObligation, obligationSweep — behind an
ObligationWiringDeps builder.

- State stays in gateway.ts: the ObligationLedger instance + durable store,
  pendingCrossTurnGate, the SHARED obligationEscalateInFlight Set (same
  reference — the concurrent-escalation guard), and all grace consts
  (injected verbatim — never-storm invariant preserved by construction).
- The escalation nudge SEND stays in gateway.ts as an injected
  sendEscalationNudge closure — the one bot.api touch remains behind the
  gateway retry-policy allowlist (new modules never call bot.api raw).
- Sweep interval stays gateway-owned (P0c: modules export the tick body).
- currentTurn read in cancelInterruptedObligation funnelled through
  getCurrentTurn() (the P9 getter funnel).
- gateway-line-ratchet lowered 26100 -> 25900 (gateway.ts now 25863 lines).

Refs #2996. Job: reference/jobs/on-the-leash.md.
