+++ scripts/gateway-line-ratchet.txt
@@ -1,1 +1,1 @@
-26100
+25900
